+++ src/Template/src/template/state/TemplateState.java
@@ -1,0 +1,12 @@
+package template.state;
+
+import GUI.Window;
+import common.state.BaseState;
+
+public class TemplateState extends BaseState {
+
+        public TemplateState() {
+                Window window = new Window(this);
+                init(window, window.getRootPanel(), Window.W, Window.H, Window.TITLE);
+        }
+}
